+++ Sphinx/ChangeLog.rst
@@ -3,6 +3,7 @@ ChangeLog
 * 02.00.00 WORKING VERSION --- November 21, 2015
 
   - Added ``ValueError`` exception if ``draw_text()`` try to draw a text containing unprintable whitespace character.
+  - Added ``draw_text_multi()`` in ``simplegui_lib_draw``.
 
   - Added alpha possibility on background color.
   - Added ``transparent`` "color" name.